Default timeout interval set to 5 min Web*Servers will not use nop 0, and even if autostart is selected, will not autostart after deepsleep dweet io cycle - it was not possible to stop it, and no longer nop 0 is used

+++ src/plugins/PowerManager.hpp
@@ -29,7 +29,7 @@ public:
 private:
   Timer *timer;
   bool wokeFromDeepSleep;
-  uint8_t timeoutIntervalS;
+  uint16_t timeoutIntervalS;
 
 
 };
